Coletor/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
urllib3==2.6.3
python-dotenv==1.2.1
streamlit==1.55.0
plotly==6.6.0
